requirements.txt
feedparser>=6.0.0
anthropic>=0.18.0
jinja2>=3.1.0
pyyaml>=6.0
python-dateutil>=2.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
